[libc++] Improved implimentation of std::cmp_less
Fixes issue #65136
In cases of integers for which wider integer types are available, all
comparisons can be made safe by converting to the wider integer.

For example, a comparision unsigned int < signed int could be
implemented through long long < long long, which is either zero
cost, or invilves a sign extention for one of the operands.

If you'd like to know more on the implimentation you can visit the
issue [#65136](https://github.com/llvm/llvm-project/issues/65136)